Removed old Bloodtalons module and renamed new module to replace it. In effect this removes the outdated 'bloodtalon overwrite' suggestion and overhauls the statistic
